% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{shapes}
\alias{shapes}
\title{Outlines: Various shapes}
\format{
A \link[Momocs2:coo_tbl]{Momocs2::coo_tbl} object with outline coordinates of some various shapes.
}
\source{
Borrowed default shapes from (c) Adobe Photoshop. Please do not send me to jail.
}
\description{
Outlines: Various shapes
}
\examples{
Momocs2::mosaic(shapes)
}
