% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{olea}
\alias{olea}
\title{Curves coordinates of olive stones}
\format{
A \code{coo_tbl} with:
\itemize{
\item \code{coo} (\code{coo_list}): olive stones, fertile halves, coordinates of two orthogonal views
\item \code{status} (\code{factor}): domesticated or wild
\item \code{accession} (\code{factor}): variety for domesticated, station for wild
\item \code{view} (\code{factor}): VL (lateral) or VD (dorsal)
}
}
\description{
Curves coordinates of olive stones
}
\section{Thanks to}{

Jean-Frédéric Terral & Sarah Ivorra - UMR ISEM, Montpellier, France.
}

\examples{
Momocs2::mosaic(olea)
}
\references{
Terral, J.‐F., Alonso, N., Capdevila, R.B.i., Chatti, N., Fabre, L., Fiorentino, G.,
Marinval, P., Jordá, G.P., Pradat, B., Rovira, N. and Alibert, P. (2004),
Historical biogeography of olive domestication (\emph{Olea europaea} L.) as revealed
by geometrical morphometry applied to biological and archaeological material.
\emph{Journal of Biogeography}, \strong{31}:63-77.
\href{https://doi.org/10.1046/j.0305-0270.2003.01019.x}{doi:10.1046/j.0305-0270.2003.01019.x}
}
\concept{curves}
