% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{shapes}
\alias{shapes}
\title{Outlines coordinates of various shapes}
\format{
A \code{coo_tbl} with:
\itemize{
\item \code{coo} (\code{coo_list}): shape coordinates
}
}
\source{
Borrowed default shapes from (c) Adobe Photoshop. Please do not send me to jail.
}
\description{
Outlines coordinates of various shapes
}
\examples{
Momocs2::mosaic(shapes)
}
\concept{outlines}
