% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{wings}
\alias{wings}
\title{Landmarks coordinates of mosquito wings}
\format{
A \link[Momocs2:coo_tbl]{Momocs2::coo_tbl} with:
\itemize{
\item \code{coo} (\code{coo_list}):coordinates
\item \code{group} (\code{factor}): grouping factor
}
}
\description{
Landmarks coordinates of mosquito wings
}
\examples{
Momocs2::mosaic(wings)
}
\references{
Rohlf, F. J. and Slice, D. (1990)
Extensions of the Procrustes method for the optimal superimposition of landmarks.
\emph{Systematic Zoology} \strong{39}:1-40.
\href{https://doi.org/10.2307/2992207}{doi:10.2307/2992207}
}
\concept{datasets}
\concept{landmarks}
